production/training-job/docker-res/requirements.txt
pandas
tqdm
torch==1.1.0
pytorch_transformers
pytorch-ignite
